Working on the app design

+++ app/src/main/java/com/example/restaurantreviewer/Model/Restaurant.kt
@@ -9,5 +9,5 @@ data class Restaurant (
     val latitude: Double,
     val longitude: Double,
     val openingHours: String,
-    var avgRating: Double? = null,
+    var avgRating: Int? = null,
         ): Serializable
